refactor: import from db-actions directly rather than from index
