+++ packages/loot-core/src/server/db/categories.js
@@ -1,7 +1,7 @@
 import { batchMessages } from '../sync';
 import { categoryModel, categoryGroupModel } from '../models';
 import { shoveSortOrders, SORT_INCREMENT } from './sort';
-import { all, delete_, first, insert, insertWithUUID, update } from './index';
+import { all, delete_, first, insert, insertWithUUID, update } from './db-actions';
 
 export async function getCategories() {
   return all(`
